+++ packages/block-editor/src/components/inserter-draggable-blocks/index.jsx
@@ -1,9 +1,12 @@
 import { Draggable } from '@wordpress/components';
-import { createBlock, store as blocksStore } from '@wordpress/blocks';
+import { store as blocksStore } from '@wordpress/blocks';
 import { useDispatch, useSelect } from '@wordpress/data';
 import { useMemo } from '@wordpress/element';
 import BlockDraggableChip from '../block-draggable/draggable-chip';
-import { INSERTER_PATTERN_TYPES } from '../inserter/block-patterns-tab/utils';
+import {
+	isPatternSynced,
+	createSyncedPatternBlock,
+} from '../inserter/block-patterns-tab/utils';
 import { store as blockEditorStore } from '../../store';
 import { unlock } from '../../lock-unlock';
 
@@ -30,11 +33,10 @@ const InserterDraggableBlocks = ( {
 	);
 
 	const patternBlock = useMemo( () => {
-		return pattern?.type === INSERTER_PATTERN_TYPES.user &&
-			pattern?.syncStatus !== 'unsynced'
-			? [ createBlock( 'core/block', { ref: pattern.id } ) ]
+		return pattern && isPatternSynced( pattern )
+			? [ createSyncedPatternBlock( pattern ) ]
 			: undefined;
-	}, [ pattern?.type, pattern?.syncStatus, pattern?.id ] );
+	}, [ pattern ] );
 
 	if ( ! isEnabled ) {
 		return children( {